fix: Fix endpoint POST /quizzes with no category
The previous validity check returned a 422 error for the valid queries when the payload did not contain a category field, or its value was None. The correct behavior is to apply no filter to the database query, as opposed to when the payload contains a category that does not exist.